backend/requirements.txt
# --- Web Framework & API ---
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1
email-validator>=2.0.0
python-multipart==0.0.20
python-dotenv==1.1.0
aiofiles==24.1.0
aiohttp>=3.9.0  # For async HTTP requests

# --- Async & Background Tasks ---
celery==5.5.0
redis==5.2.1

# --- Database ---
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
alembic==1.13.2
pgvector>=0.2.0
uuid6>=1.0.3  # For UUID v7 support (ordered, timestamp-based)

# --- Document Processing (Simplified) ---
docling>=2.80.0
pypdf>=4.0.0
PyMuPDF>=1.24.0
pillow>=11.0.0

# --- LLM Integrations ---
groq>=0.12.0
google-generativeai>=0.8.3
openai>=1.60.0
anthropic>=0.42.0
google-cloud-aiplatform>=1.70.0
requests==2.32.3
httpx>=0.27.0

# --- Authentication & Security ---
python-jose==3.3.0
passlib==1.7.4
cryptography>=42.0.0
PyJWT>=2.12.0

# --- Utilities ---
numpy>=1.26.0
scikit-learn>=1.4.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0  # Requerido por sentence-transformers (versión CPU)
sentence-transformers>=3.0.0  # Para embeddings locales sin APIs
python-json-logger==2.0.7
structlog>=24.0.0
pydantic-extra-types>=2.0.0
tenacity>=8.0.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

